refactor(catalog): move helper functions into helpers file. delete redundent usePlugins hook
