+++ public/app/features/plugins/admin/helpers.ts
@@ -1,5 +1,80 @@
 import { config } from '@grafana/runtime';
+import { CatalogPlugin, LocalPlugin, Plugin } from './types';
 
 export function isGrafanaAdmin(): boolean {
   return config.bootData.user.isGrafanaAdmin;
+}
+
+export function mapRemoteToCatalog(plugin: Plugin): CatalogPlugin {
+  const {
+    name,
+    slug: id,
+    description,
+    version,
+    orgName,
+    popularity,
+    downloads,
+    updatedAt,
+    createdAt: publishedAt,
+  } = plugin;
+  const catalogPlugin = {
+    description,
+    downloads,
+    id,
+    info: {
+      logos: {
+        small: `https://grafana.com/api/plugins/${id}/versions/${version}/logos/small`,
+        large: `https://grafana.com/api/plugins/${id}/versions/${version}/logos/large`,
+      },
+    },
+    name,
+    orgName,
+    popularity,
+    publishedAt,
+    updatedAt,
+    version,
+    isInstalled: false,
+  };
+  return catalogPlugin;
+}
+
+export function mapLocalToCatalog(plugin: LocalPlugin): CatalogPlugin {
+  const {
+    name,
+    info: { description, version, logos, updated, author },
+    id,
+  } = plugin;
+  return {
+    description,
+    downloads: 0,
+    id,
+    info: { logos },
+    name,
+    orgName: author.name,
+    popularity: 0,
+    publishedAt: '',
+    updatedAt: updated,
+    version,
+    isInstalled: true,
+  };
+}
+
+export function applySearchFilter(searchBy: string | undefined, plugins: CatalogPlugin[]): CatalogPlugin[] {
+  if (!searchBy) {
+    return plugins;
+  }
+
+  return plugins.filter((plugin) => {
+    const fields: String[] = [];
+
+    if (plugin.name) {
+      fields.push(plugin.name.toLowerCase());
+    }
+
+    if (plugin.orgName) {
+      fields.push(plugin.orgName.toLowerCase());
+    }
+
+    return fields.some((f) => f.includes(searchBy.toLowerCase()));
+  });
 }
